+++ raw/2_sub.sh
@@ -1,15 +1,33 @@
 #!/bin/bash
+test `whoami` = "root" || exit 0
 hostname
-grep -e no_proxy /etc/environment || echo -e 'http_proxy=http://vb:55555\nhttps_proxy=http://vb:55555\nno_proxy=127.0.0.1,vb,va,vd,*.internal,localhost,*.testnet.tech' >> /etc/environment
-rpm -i ~uuu/*.rpm
-( cd / ; tar -xpf ~uuu/yum.tar )
+localip=`ip addr show eth0|grep -Po 'inet ([0-9.]*)'|grep -Po '[0-9.]*'|head -n 1`
+cat >/dev/null <<EON
+sysctl net.ipv6.conf.all.disable_ipv6=0
+sysctl -w net.ipv6.conf.all.disable_ipv6=0
+sed -i 's/keepcache=0/keepcache=1/g' /etc/yum.conf
+cat <<EOI >> /usr/share/doc/glibc-common-2.17/gai.conf
+precedence ::ffff:0:0/96  100
+precedence  ::1/128       50
+precedence  ::/0          40
+precedence  2002::/16     30
+precedence ::/96          20
+EOI
+EON
+
+cd /home/uuu
+#rm -f squid*
+#grep -e no_proxy /etc/environment || echo -e 'http_proxy=http://vb:55558\nhttps_proxy=http://vb:55558\nno_proxy=127.0.0.1,vb,va,vd,*.internal,localhost,*.testnet.tech' >> /etc/environment
+ls -1 *.rpm | xargs yum -y localinstall
+test -f /usr/bin/consul || exit 1
+#( cd / ; tar -xpf /home/uuu/yum.tar )
 consul -autocomplete-install
 echo '{"server": true,"data_dir": "/var/lib/consul","log_level": "INFO","ui_config":{"enabled":true},"rejoin_after_leave": true,"leave_on_terminate": false,"retry_join": ["va","vb","vd"],"bootstrap": false, "bootstrap_expect": 3 }' > /etc/consul.d/consul.json
 chown consul.consul /etc/consul.d/consul.json
 sudo systemctl enable --now consul
 
-mv ~uuu/consul-template /usr/bin/
-cp {~uuu,/etc/haproxy}/haproxy.conf.ctmpl
+mv consul-template /usr/bin/
+#cp {/home/uuu,/etc/haproxy}/haproxy.conf.ctmpl
 timedatectl set-timezone Europe/Moscow
 #yum -y install https://download.postgresql.org/pub/repos/yum/reporpms/EL-7-x86_64/pgdg-redhat-repo-latest.noarch.rpm
 sed -i "s/en_US/ru_RU/g" /etc/locale.conf
@@ -21,13 +39,14 @@ LC_ALL=ru_RU.UTF8
 PGDATA=/var/lib/pgsql/13/data
 PATH="/usr/pgsql-13/bin:$PATH"
 EOC
-
 disablereps="--disablerepo=pgdg15 --disablerepo=pgdg14 --disablerepo=pgdg12"
-yum install -y postgresql13-server postgres13-agent haproxy epel-release python-psycopg2 chrony wget zip unzip jq vim curl lsof git $disablereps
-yum install -y patroni-consul $disablereps # postgresql13-plpython3
-
-cp {~uuu/,/etc/patroni/}patroni.yml
+#yum install -C -y epel-release deltarpm $disablereps || yum install -y epel-release deltarpm $disablereps
+#yum install -C -y postgresql13-server postgres13-agent haproxy epel-release python-psycopg2 chrony wget zip unzip jq vim curl lsof git keepalived patroni-consul $disablereps || yum install -y postgresql13-server postgres13-agent haproxy epel-release python-psycopg2 chrony wget zip unzip jq vim curl lsof git keepalived patroni-consul $disablereps
+test -f /usr/bin/zip || exit 1
+sed "s/IPADDR/$localip/g;s/HOSTNAME/`hostname`/g" /home/uuu/keepalived.conf.0 > /etc/keepalived/keepalived.conf
+cp {/home/uuu/,/etc/patroni/}patroni.yml
 chown postgres.postgres /etc/patroni/patroni.yml
-localip=`ip addr show eth0|grep -Po 'inet ([0-9.]*)'|grep -Po '[0-9.]*'`
-sed -i "s/IPADDR/$localip/g;s/HOSTNAME/`hostname`/g"
+sed "s/IPADDR/$localip/g;s/HOSTNAME/`hostname`/g" /home/uuu/patroni.yml > /etc/patroni/patroni.yml
+cp /home/uuu/haproxy.cfg /etc/haproxy/
+cat /etc/keepalived/keepalived.conf /etc/patroni/patroni.yml /etc/consul.d/consul.json /etc/haproxy/haproxy.cfg /etc/environment
